Starting work on node lookup and validation
